Número de conexões limitadas

+++ src/main/java/br/com/unioeste/oficinamecanicapg/core/repositories/BairroRepository.java
@@ -3,5 +3,5 @@ package br.com.unioeste.oficinamecanicapg.core.repositories;
 import br.com.unioeste.oficinamecanicapg.core.entities.Bairro;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BairroRepository extends JpaRepository<Bairro, Long>{
+public interface BairroRepository extends JpaRepository<Bairro, Integer>{
 }
